LayoutTests/imported/w3c/web-platform-tests/html/browsers/sandboxing/sandbox-disallow-scripts-via-unsandboxed-popup.tentative-expected.txt

CONSOLE MESSAGE: Blocked script execution in 'about:srcdoc' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.
CONSOLE MESSAGE: TypeError: null is not an object (evaluating 'window.opener.top')


Harness Error (TIMEOUT), message = null

TIMEOUT Sandboxed => unsandboxed popup Test timed out
